requirements.txt
annotated-types==0.7.0
colorama==0.4.6
numpy==2.2.6
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
